Starting the API stuff
This is only the beginning of a legend.

Basically it's the barebones start of me structuring stuff. This API
will be similar in many ways to bukkit, because I liked using bukkit a
lot. RIP by the way.

Server will come a bit soon, once more of this is done and I'll run some
network tests before implementing a proper client into the RozWorld
build.