Adding world and explosion stuff, some server stuff
A bit of a mix of stuff here.

Lots in the Level folder to do with the world, including the world
itself, chunks and tiles (walls and floors too).

Some additions to the server class for handling these new classes (I
still have to make sure everything is done correctly though).

Explosion class has been added, the Intensity value is how the damage is
distributed throughout the explosion. There'll be an equation to
calculate damage based on that variable later on when the server is
implemented anyway.