Redoing the API from scratch, read desc...
THIS COMMIT IS NOT TO BE USED... WILL BE SETTING UP OTHER TREES, MASTER
WILL NOT BE USED RIGHT NOW

The old API and stuff was all done out of sync, it would've taken some
mishmashing to get it to all work together. This 'recode' is taking a
better approach, stuff'll be defined in the API first, then implemented
in the server library so that extensive testing can go on the whole way.

The RozWorld client itself will get a redo too in due time, the server
will be tested first however :neckbeard:

Other news to do with the site, roz.world will be sorted out soon since
I moved hosts (which is why oddmatics.uk is also down). Soon.jpg