Change player message theory
Rather than sending a message to a player posing as another player, it
is better to have the player send the message to another player.

Player.SendPrivateMessageTo() will send a private message from the
Player instance the method was called from, to the Player instance in
recipient. Makes more sense that way.

The only one that hasn't changed is Player.SendMessage(), which will
send a generic unformatted message to that Player instance only (hence
the missing 'to' in SendMessage as opposed to 'SendInviteTo').