Add chat hooking to ICommandCaller
This will allow plugins/commands to hook a caller's chat so that the
next message or so will be sent immediately to the delegate rather than
processed by the server.

For example, say you had a command /kill that killed a player but asked
for confirmation before doing so, in which a player would send 'yes'
after sending /kill. The command handler for /kill would call
HookChatToCallback() on the sender of /kill, in which the next delegate
checks if they confirmed or denied the /kill and then releases the hook
at the end of processing. :neckbeard:

Hopefully that made sense, this is all to simplify things for plugin
authors!